requirements.txt
aiofiles==0.7.0
aioredis==1.3.1
aiosqlite==0.17.0
anyio==3.3.2
arrow==1.1.1
asgiref==3.4.1
astroid==2.8.0
async-timeout==3.0.1
automium==0.2.6
automium-web==0.1.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
Chameleon==3.9.1
chardet==4.0.0
charset-normalizer==2.0.5
click==8.0.1
cookiecutter==1.7.3
cryptography==35.0.0
databases==0.5.2
distlib==0.3.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
fastapi-admin==1.0.3
fastapi-cache==0.1.0
fastapi-caching==0.3.0
fastapi-chameleon==0.1.11
fastapi-users==7.0.0
fastapi-users-db-sqlalchemy==1.0.0
filelock==3.0.12
Flask==2.0.2
greenlet==1.1.1
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.19.0
idna==3.2
iso8601==0.1.16
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
jinja2-time==0.2.0
lazy-object-proxy==1.6.0
legacy==0.1.6
makefun==1.11.3
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.28
Padding==0.5
passlib==1.7.4
pendulum==2.1.2
pipenv==2021.5.29
platformdirs==2.3.0
poyo==0.5.0
protobuf==3.19.3
pycparser==2.20
pycryptodome==3.10.4
pydantic==1.8.2
PyJWT==2.1.0
pylint==2.11.1
pyOpenSSL==21.0.0
pypika-tortoise==0.1.1
pypng==0.0.21
PyQRCode==1.2.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
python-slugify==5.0.2
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.23
starlette==0.14.2
structlog==21.1.0
text-unidecode==1.3
toml==0.10.2
tortoise-orm==0.17.7
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
watchgod==0.7
websockets==10.0
Werkzeug==2.0.2
wrapt==1.12.1
